util_classify

Classify a raster into proportions based upon a vector of class weightings.

util_classify(x, weighting, level_names = NULL)

Arguments

x
2D matrix

weighting
Vector of numeric values.

level_names
Vector of names for the factor levels.

Value

RasterLayer

Details

The number of elements in the weighting vector determines the number of classes in the resulting matrix. The classes start with the value 1. If non-numerical levels are required, the user can specify a vector to turn the numerical factors into other data types, for example into character strings (i.e. class labels). If the numerical vector of weightings does not sum up to 1, the sum of the weightings is divided by the number of elements in the weightings vector and this is then used for the classification.

Examples

weight <- c(0.5, 0.25, 0.25)
util_classify(fbmmap, weight,
              level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))

#> class       : RasterLayer 
#> dimensions  : 200, 200, 40000  (nrow, ncol, ncell)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 200, 0, 200  (xmin, xmax, ymin, ymax)
#> coord. ref. : NA 
#> data source : in memory
#> names       : layer 
#> values      : 0, 1  (min, max)
#> attributes  :
#>  ID Categories
#>   0 Land Use 1
#>   1 Land Use 2
#>   2 Land Use 3
#>